chunk4: relocate analysis types, harden invariants, extract adapter logic
Domain boundary cleanup based on 10-subagent domain-modeling review:

- Move AnalysisWindow from domain/ to analysis/ (not a domain concept;
  only consumed by phase_segmentation and summary_builder). DrivingPhase
  enum stays in domain/ as the genuine domain concept.

- Move PhaseEvidence TypedDict from domain/finding.py to analysis/_types.py
  (pipeline serialization concern, not domain). Inline cruise_fraction
  parsing in Finding.from_payload() to avoid cross-layer dependency.

- Extract Report.from_summary() to report/mapping.py as
  build_report_from_summary() (adapter logic parsing JSON dicts does not
  belong on a pure domain object).

- Harden Car invariant: reject zero tire dimensions (val <= 0), aligning
  Car.__post_init__ with TireSpec.from_aspects.

- Harden SpeedSource invariant: MANUAL kind requires manual_speed_kmh > 0
  (was only checking is not None).

- Add Finding REF_ prefix cross-check: warn when explicit kind conflicts
  with REF_ finding_id prefix convention.

- Remove unused type: ignore comment in Finding.from_payload().

- Update docs/domain-model.md and docs/ai/repo-map.md to reflect all
  relocations and invariant changes.

- Update tests: new import paths, tightened invariant expectations,
  5 new targeted invariant tests.
